Create context package
