+++ internal/laputa/middleware.go
@@ -36,10 +36,3 @@ func (l *laputa) LogHandler() echo.MiddlewareFunc {
 		}
 	}
 }
-
-func JSTFormatter(key string) zap.TimeFormatter {
-	return zap.TimeFormatter(func(t time.Time) zap.Field {
-		jst := time.FixedZone("Asia/Tokyo", 9*3600)
-		return zap.String(key, t.In(jst).Format(time.ANSIC))
-	})
-}
